src/faketime: implicit '-f' for $offset that is faketime-known relative
When $offset arg looks like the relative value described in README
section "4d) Using offsets for relative dates", set USEDIRECT variable
even when '-f' option is not present. Without the '-f' the $offset would
be useless so this is convenience feature for the user. Using '-f' in
command line to explicitly set USEDIRECT is retained.