Only let the owner clean up semaphore and shared memory
I noticed a bug in the semaphore handling, when using the System V semaphore
backend:

  $ LD_PRELOAD=./src/libfaketime.so.1 bash -c "echo foo | sed s/foo/bar/"
  libfaketime: In lock_for_stat(), ft_sem_lock failed: Invalid argument
  [...exited with error...]

(Beware, the above command-line is not 100% deterministic; sometimes it
succeeds.)

Looking at the strace for the above command-line, it seems the bash echo
builtin process (or thread?) decides to remove the semaphore upon
exiting, while it's still in use by the sed process. sed then gets
EINVAL error ("Invalid argument") on its next semop call.

The root cause is a semantic difference between POSIX sem_unlink and
SysV semop(..., IPC_RMID), the two implementations for ft_sem_unlink:

* sem_unlink allows the semaphore to be used afterwards, as long as a
  process has a reference to the semaphore.
* semop(..., IPC_RMID) removes the semaphore immediately, and further
  use results in EINVAL error.

AFAICT, the simplest fix is to only let the owner of the semaphore (and
shared memory) do the clean up, which is what this patch does. Both
semaphore backends pass the tests with this change.